# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestão Escolar",
        "description": "Sistema de Gestão Escolar",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/alunos": {
            "get": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Listar Alunos",
                "operationId": "alunos/index",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "order  accepts 'asc' or 'desc'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/aluno": {
            "post": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Cadastrar Aluno",
                "operationId": "aluno/store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "telefone",
                                    "genero"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Aluno Teste"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "teste@teste.com"
                                    },
                                    "telefone": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "99999999999"
                                    },
                                    "data_nascimento": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1111-11-11"
                                    },
                                    "genero": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Masculino"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/aluno/{id}": {
            "get": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Pesquisar Aluno",
                "operationId": "aluno/show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Aluno",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Editar Aluno",
                "operationId": "aluno/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Aluno",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "telefone",
                                    "genero"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Aluno Teste"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "teste@teste.com"
                                    },
                                    "telefone": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "99999999999"
                                    },
                                    "data_nascimento": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1111-11-11"
                                    },
                                    "genero": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Masculino"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Excluir Aluno",
                "operationId": "aluno/destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Aluno",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}